@@ -135,9 +135,15 @@ ENDIF(SLEPC_FOUND)
 #   SET(HAVE_LIBMESH 1)
 # ENDIF(LIBMESH_FOUND)
 
+find_package(Eigen3 REQUIRED NO_MODULE)
+
 SET(EIGEN_SAFE_TO_USE_STANDARD_ASSERT_MACRO FALSE)
 
-set(CMAKE_CXX_STANDARD 11)
+#find_package(Ceres NO_MODULE REQUIRED)
+
+#find_package(glog NO_MODULE REQUIRED)
+
+set(CMAKE_CXX_STANDARD 14)
 
 #############################################################################################
 ### Build options
