@@ -26,7 +26,7 @@ double InitalValueU(const std::vector < double >& x) {
   for(unsigned k = 0; k < x.size(); k++) {
     value +=  x[k] * x[k]; //consistency
 //     value +=  x[k] * x[k] * x[k]; //cubic
-//     value +=  x[k] * x[k] * x[k] * x[k];//quartic
+//    value +=  x[k] * x[k] * x[k] * x[k];//quartic
   }
 
 
@@ -45,14 +45,14 @@ bool SetBoundaryCondition(const std::vector < double >& x, const char SolName[],
   for(unsigned k = 0; k < x.size(); k++) {
     value +=  x[k] * x[k]; //consistency
 //     value +=  x[k] * x[k] * x[k]; //cubic
-//     value +=  x[k] * x[k] * x[k] * x[k];//quartic
+//    value +=  x[k] * x[k] * x[k] * x[k];//quartic
   }
   
   return dirichlet;
 }
 
-// unsigned numberOfUniformLevels = 2; //consistency
-unsigned numberOfUniformLevels = 2; //cubic-quartic 2->6 //cubic Marta4Quad Tri Mix
+ unsigned numberOfUniformLevels = 1; //consistency
+//unsigned numberOfUniformLevels = 4; //cubic-quartic 2->6 //cubic Marta4Quad Tri Mix
 //unsigned numberOfUniformLevels = 2; //cubic-quartic 2->4 mappa a 4->6 //cubic Marta4Fine
 
 
@@ -73,14 +73,14 @@ int main(int argc, char** argv) {
 
 
 
-  //char fileName[100] = "../input/martaTest4.neu"; // good form 2->6 in serial but in parallel use martaTest4Fine
-  //char fileName[100] = "../input/martaTest4Fine.neu"; // works till 144 nprocs +2
+//   char fileName[100] = "../input/martaTest4.neu"; // good form 2->6 in serial but in parallel use martaTest4Fine
+//   char fileName[100] = "../input/martaTest4Fine.neu"; // works till 144 nprocs +2
 //   char fileName[100] = "../input/martaTest4Finer.neu"; // works till 144 nprocs +4
-  //char fileName[100] = "../input/martaTest4Tri.neu";
-  //char fileName[100] = "../input/martaTest4Unstr.neu"; // works till 144 nprocs
+//   char fileName[100] = "../input/martaTest4Tri.neu";
+//   char fileName[100] = "../input/martaTest4Unstr.neu"; // works till 144 nprocs
 //   char fileName[100] = "../input/martaTest4-3D.neu"; // works till 288 nprocs 0.2
-  //char fileName[100] = "../input/martaTest4-3Dfine.neu"; // works till 576 and more nprocs +1 0.1
-  char fileName[100] = "../input/martaTest4-3D-tet.neu"; // works till 288 nprocs 0.2
+//  char fileName[100] = "../input/martaTest4-3Dfine.neu"; // works till 576 and more nprocs +1 0.1
+   char fileName[100] = "../input/martaTest4-3D-tet.neu"; // works till 288 nprocs 0.2
 
   mlMsh.ReadCoarseMesh(fileName, "fifth", scalingFactor);
   MPI_Barrier(MPI_COMM_WORLD);
@@ -375,9 +375,9 @@ void GetL2Norm(MultiLevelSolution & mlSol, MultiLevelSolution & mlSolFine) {
 
       double soluExact_gss = 0.;
       for(unsigned k = 0; k < dim; k++) {
-        soluExact_gss += xg[k] * xg[k];//consistency
+         soluExact_gss += xg[k] * xg[k];//consistency
 //         soluExact_gss += xg[k] * xg[k] * xg[k]; // cubic
-//         soluExact_gss += xg[k] * xg[k] * xg[k] * xg[k];// quartic
+//        soluExact_gss += xg[k] * xg[k] * xg[k] * xg[k];// quartic
       }
 
       error_solExact_norm2 += (soluNonLoc_gss - soluExact_gss) * (soluNonLoc_gss - soluExact_gss) * weight;
